2.7
{
    "filename": "links-2.7.tar.gz",
    "md5": "6b5ca02e180aed2f78ca2712f85e3ba5",
    "sha1": "e0773f2b23397bcbd08d5a3145d94e446dfb4969",
    "sha256": "2ff4f4bfa5ee1d11d5e2eed39311824d65aa0610da63b22bf4b14793c8f48aa5"
}

+++ png.c
@@ -157,7 +157,7 @@ static void png_row_callback(png_structp png_ptr, png_bytep new_row, png_uint_32
 				*row_num), new_row, cimg->width
 				*channels);
 		}else{
-			if ((unsigned)cimg->width > MAXINT / 2 / channels) overalloc();
+			if ((unsigned)cimg->width > (unsigned)MAXINT / 2 / channels) overalloc();
 			tmp=mem_alloc(cimg->width*2*channels);
 			a2char_from_unsigned_short(tmp, (unsigned short *)(cimg->buffer
 				+cimg->buffer_bytes_per_pixel
